migrate from deprecated snowpack to vite
