+++ .github/workflows/build.yaml
@@ -10,16 +10,9 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-
-      - name: Setup Node
-        uses: actions/setup-node@v3
+      - uses: actions/setup-node@v3
         with:
-          node-version: '16'
-          cache: "npm"
-          registry-url: https://npm.pkg.github.com/
-
+          node-version: 16
+          cache: 'npm'
       - run: npm ci
-        env:
-          NODE_AUTH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-      - run: npx snowpack build
+      - run: npm run build
