--- Makefile
@@ -1,6 +1,0 @@
-test:
-	npx snowpack
-
-build:
-	npx snowpack build
-
