--- assets/sass/fontawesome.scss
@@ -1,7 +1,0 @@
-@charset "utf-8";
-
-$fa-font-path: "webfonts";
-@import "./scss/fontawesome.scss";
-@import "./scss/solid.scss";
-@import "./scss/regular.scss";
-@import "./scss/brands.scss";
