+++ js/common/utilities.js
