+++ js/workflows/ble.js
@@ -2,8 +2,9 @@
  * This class will encapsulate all of the workflow functions specific to BLE
  */
 
+import {FileTransferClient} from 'ble-file-transfer-js';
+
 import {CONNTYPE, CONNSTATE} from '../constants.js';
-import {FileTransferClient} from 'https://cdn.jsdelivr.net/gh/adafruit/ble-file-transfer-js@1.0.2/adafruit-ble-file-transfer.js';
 import {Workflow} from './workflow.js';
 import {GenericModal} from '../common/dialogs.js';
 import {sleep, getUrlParam} from '../common/utilities.js';
