+++ public/assets/js/device.js
@@ -1,3 +1,5 @@
+// This file is needed and loaded by https://github.com/adafruit/circuitpython/blob/main/supervisor/shared/web_workflow/static/code.html
+
 // Boot strap load everything from code.circuitpython.org
 let SITE = "https://code.circuitpython.org";
 if (location.hostname == "localhost" || location.hostname == "127.0.0.1") {
