+++ sass/layout/_header_mobile.scss
