+++ sass/pages/_editor.scss
