+++ sass/style.scss
@@ -1,5 +1,8 @@
 @charset "utf-8";
 
+@import '/node_modules/@fortawesome/fontawesome-free/css/all.min.css';
+@import '/node_modules/xterm/css/xterm.css';
+
 @import './vendors/normalize';
 
 @import './base/variables';
